requirements.txt
asgiref==3.4.1
Django==3.2.8
django-crum==0.7.9
django-widget-tweaks==1.4.8
Pillow==8.3.2
python-decouple==3.5
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
